Working gemv kernel
